fix(analysis): Compare profile requests so that we re-request the origin if the routing params have
Previously e.g. changing the date would have no effect because the profile request was cached at the

time the scenario was loaded. Now we use the new profile request and compare to the existing one to

determine if a reload of the origin is necessary, and we pass in the correct request.