feat(project): Add user feedback when downloading street network and census data
It takes up to a few minutes to download Census data and street network data. This shows the user
the status of that process, and also ensures that all indicators are loaded into the UI once they
have been fetched, so that they are available for use in analysis.

fixes #339 and #323 and fixes #322