Refactoring opportunity dataset docs for clarity (#950)
* Refactoring opportunity dataset docs for clarity

This is mostly a reorganization of existing content. My goal was to put the more important information first before trailing off into the necessary specifics of implementation and formatting. Hopefully the example CSV better clarifies how these should be formatted, though most of the requirements for that are still given in the text. 

I removed the note about later LODES data as it seems only 2015 data are available at the moment.

* Cleaning up LODES section mostly

* docs(lodes): clarify user manual

Co-authored-by: Anson Stewart <ansonstewart@gmail.com>