Use Auth0 as recommended + statically render
This commit updates to use the nextjs-auth0 integration made by Auth0. It does all the verification on the server side and delivers a cookie to the browser that contains the session data so that no client side libraries or verification is needed for Authentication.

This commit also removes the need to server side render everything by removing `getInitialProps` in the `pages/_app` component. That caused a cascade of changes including moving to a Layout approach for pages and creating higher order page components including with-auth and with-redux.